requirements.txt
Pillow==3.3.1
numpy==1.11.1
pypiwin32==219
